See this thread of discussion to understand what I did in regex.rs and how that static thing works ([1]).
In short, everything works as expected now. There is a String that is calculated
only once during the program lifetime. This allows for very interesting delegation
from Wrapper types to Core types. In this case, form CharGroup regex type to the
RegexLexeme (RegexLexemes that are also CharGroups can be wrapped in '[...]' block
to be exported to other regular expressions). Pay attention, there is literally no
bloat here, CharGroup is defined only ONCE. To use this functionality, you only
need to mark that your custom type implements CharGroup.

Also attaching this [2] for reference

[1] https://users.rust-lang.org/t/lazy-static-macro-does-not-compile-when-using-self-type-parameter/44917/4?u=gevorgyana
[2] https://users.rust-lang.org/t/how-to-avoid-recalculating-a-formatted-string-at-runtime/44895/8